Add exclude list

+++ src/index.ts
@@ -3,14 +3,35 @@ import dotenv from 'dotenv';
 import { Sonarr } from './sonarr';
 import { Trakt } from './trakt';
 import { sendDiscordMessage } from './utils';
+import * as fs from 'fs';
 
 dotenv.config();
 
+function getExludes() {
+    let ret: string[] = [];
+    if (fs.existsSync('./exclude.txt'))
+        ret = fs.readFileSync('./exclude.txt').toString().split(/\r?\n/g);
+    if (ret[ret.length - 1] === '') ret = ret.slice(0, -1);
+    return ret;
+}
+
+function isSeriesExcluded(series, excludes) {
+    const slug = series.titleSlug.replace(/[^a-zA-Z]/g, '');
+    for (let exclude of excludes) {
+        exclude = exclude.replace(/[^a-zA-Z]/g, '');
+        if (slug.match(new RegExp(exclude, 'gi')) !== null) return true;
+    }
+    return false;
+}
+
 async function main() {
     try {
+        const excludes = getExludes();
+
         const trakt = new Trakt();
         await trakt.authorize();
         const new_watched = await trakt.newWatched();
+        if (!new_watched.length) return;
 
         const sonarr = new Sonarr();
         const allSeries = await sonarr.getAllSeries();
@@ -21,6 +42,7 @@ async function main() {
             const episode = episodes.data.find(e => e.tvdbId === watched.episode.ids.tvdb);
             const episodeFileId: number = episode.episodeFileId;
             if (episodeFileId === 0) continue;
+            if (isSeriesExcluded(series, excludes)) continue;
 
             try {
                 await sonarr.deleteFile(episode.episodeFileId);
